Move bgfx shader construction into ShaderCompiler class
