+++ Plugins/NativeEngine/Source/NativeEngine.h
@@ -283,13 +283,6 @@ namespace Babylon
         uint16_t m_nextId{0};
     };
 
-    struct UniformInfo final
-    {
-        uint8_t Stage{};
-        // uninitilized bgfx resource is BGFX_INVALID_HANDLE. 0 can be a valid handle.
-        bgfx::UniformHandle Handle{bgfx::kInvalidHandle};
-    };
-
     struct TextureData final
     {
         ~TextureData()
@@ -319,6 +312,12 @@ namespace Babylon
         std::unique_ptr<bimg::ImageContainer> Image;
     };
 
+    struct UniformInfo final
+    {
+        uint8_t Stage{};
+        bgfx::UniformHandle Handle{bgfx::kInvalidHandle};
+    };
+
     struct ProgramData final
     {
         ProgramData() = default;
@@ -330,9 +329,9 @@ namespace Babylon
             bgfx::destroy(Program);
         }
 
-        std::unordered_map<std::string, uint32_t> AttributeLocations{};
-        std::unordered_map<std::string, UniformInfo> VertexUniformNameToInfo{};
-        std::unordered_map<std::string, UniformInfo> FragmentUniformNameToInfo{};
+        std::unordered_map<std::string, uint32_t> VertexAttributeLocations{};
+        std::unordered_map<std::string, UniformInfo> VertexUniformInfos{};
+        std::unordered_map<std::string, UniformInfo> FragmentUniformInfos{};
 
         bgfx::ProgramHandle Program{};
