+++ Plugins/NativeEngine/Source/ShaderCompilerTraversers.cpp
@@ -12,24 +12,20 @@
 
 #include <stdexcept>
 
+using namespace glslang;
+
 namespace Babylon::ShaderCompilerTraversers
 {
-    /// <summary>
     /// The purpose of everything in this namespace is to modify the glslang abstract
     /// syntax tree generated by parsing Babylon.js shaders so that those shaders
     /// can be recompiled to target native shader languages such as DirectX, OpenGL,
     /// and Metal.
-    /// </summary>
     namespace
     {
-        using namespace glslang;
-
-        /// <summary>
         /// Helper method to replace symbols in a glslang AST. This operation is done
         /// by several of the traversers in this file.
-        /// </summary>
-        /// <param name="nameToReplacement">Map from symbol names to the node which should replace that symbol.</param>
-        /// <param name="symbolToParent">Vector of symbols to be replaced along with their parents in the AST.</param>
+        /// @param nameToReplacement Map from symbol names to the node which should replace that symbol.
+        /// @param symbolToParent Vector of symbols to be replaced along with their parents in the AST.
         void makeReplacements(
             std::map<std::string, TIntermTyped*> nameToReplacement,
             std::vector<std::pair<TIntermSymbol*, TIntermNode*>> symbolToParent)
@@ -74,39 +70,36 @@ namespace Babylon::ShaderCompilerTraversers
             }
         }
 
-        /// <summary>
         /// Helper method to determine whether an element in the AST is a linker object,
         /// which is a special part of the AST used to enumerate symbols for linking.
-        /// </summary>
-        /// <param name="path">The path to the element in question.</param>
-        /// <returns>True if path is for a linker object, false otherwise.</returns>
+        /// @param path The path to the element in question.
+        /// @returns True if path is for a linker object, false otherwise.
         bool isLinkerObject(const TIntermSequence& path)
         {
             auto* agg = path.size() > 1 ? path[1]->getAsAggregate() : nullptr;
             return agg && agg->getOp() == EOpLinkerObjects;
         }
 
-        /// <summary>
         /// This traverser collects all non-sampler uniforms and creates a new struct
         /// called "Frame" to contain them. This is necessary to correctly transpile
         /// for DirectX and Metal.
-        /// </summary>
         class NonSamplerUniformToStructTraverser final : private TIntermTraverser
         {
         public:
+            static ScopeT Traverse(TProgram& program, IdGenerator& ids)
+            {
+                auto* scope = new AllocationsScope();
+                Traverse(program.getIntermediate(EShLangVertex), ids, *scope);
+                Traverse(program.getIntermediate(EShLangFragment), ids, *scope);
+                return std::unique_ptr<AllocationsScopeBase>(scope);
+            }
 
-            /// <summary>
+        private:
             /// This scope contains the new allocations added to the AST to represent
             /// the new struct, information about the elements it contains, dereferencing
             /// operations, etc.
-            /// </summary>
             class AllocationsScope : AllocationsScopeBase
             {
-            public:
-                ~AllocationsScope() override
-                {
-                }
-            
             private:
                 friend NonSamplerUniformToStructTraverser;
                 std::vector<std::unique_ptr<TType>> Types{};
@@ -114,11 +107,7 @@ namespace Babylon::ShaderCompilerTraversers
                 std::vector<std::unique_ptr<TArraySizes>> ArraySizes{};
             };
 
-            ~NonSamplerUniformToStructTraverser() override
-            {
-            }
-
-            void visitSymbol(TIntermSymbol* symbol) override
+            virtual void visitSymbol(TIntermSymbol* symbol) override
             {
                 // Collect all non-sampler uniforms and add the to the list of elements to process.
                 if (symbol->getType().getQualifier().isUniformOrBuffer() && symbol->getType().getBasicType() != EbtSampler)
@@ -139,16 +128,7 @@ namespace Babylon::ShaderCompilerTraversers
                 }
             }
 
-            static ScopeT Traverse(TProgram& program, IdGenerator& ids)
-            {
-                auto* scope = new AllocationsScope();
-                Traverse(program.getIntermediate(EShLangVertex), ids, *scope);
-                Traverse(program.getIntermediate(EShLangFragment), ids, *scope);
-                return std::unique_ptr<AllocationsScopeBase>(scope);
-            }
-
-        private:
-            static void Traverse(glslang::TIntermediate* intermediate, IdGenerator& ids, AllocationsScope& scope)
+            static void Traverse(TIntermediate* intermediate, IdGenerator& ids, AllocationsScope& scope)
             {
                 NonSamplerUniformToStructTraverser traverser{};
                 intermediate->getTreeRoot()->traverse(&traverser);
@@ -271,36 +251,37 @@ namespace Babylon::ShaderCompilerTraversers
             std::vector<std::pair<TIntermSymbol*, TIntermNode*>> m_symbolsToParents{};
         };
 
-        /// <summary>
         /// Changes the types of all float, vec2, and vec3 uniforms to vec4. This is required
         /// for OpenGL and Metal.
-        /// </summary>
         class UniformTypeChangeTraverser final : private TIntermTraverser
         {
         public:
+            static ScopeT Traverse(TProgram& program, IdGenerator& ids)
+            {
+                auto* scope = new AllocationsScope();
+                Traverse(program.getIntermediate(EShLangVertex), ids, *scope);
+                Traverse(program.getIntermediate(EShLangFragment), ids, *scope);
+                return std::unique_ptr<AllocationsScopeBase>(scope);
+            }
+
+        private:
             class AllocationsScope : AllocationsScopeBase
             {
-            public:
-                ~AllocationsScope() override
-                {
-                }
-
             private:
                 friend UniformTypeChangeTraverser;
                 std::vector<std::unique_ptr<TArraySizes>> ArraySizes{};
             };
 
-            ~UniformTypeChangeTraverser() override
+            UniformTypeChangeTraverser(TIntermediate* intermediate, AllocationsScope& scope)
+                : m_intermediate{intermediate}
+                , m_scope{scope}
             {
             }
 
-            /// <summary>
             /// Because no accumulation or cross-correlation is necessary for this change
             /// (i.e. each operation acts only on a single symbol), we do all the work inside
             /// the traverser itself.
-            /// </summary>
-            /// <param name="symbol"></param>
-            void visitSymbol(TIntermSymbol* symbol) override
+            virtual void visitSymbol(TIntermSymbol* symbol) override
             {
                 auto& type = symbol->getType();
 
@@ -312,7 +293,7 @@ namespace Babylon::ShaderCompilerTraversers
                     // something like (vec3.xyz).x. I suspect this is unlikely to cause problems.
 
                     auto* oldType = type.clone();
-                    
+
                     TPublicType publicType{};
                     publicType.qualifier = type.getQualifier();
 
@@ -422,23 +403,7 @@ namespace Babylon::ShaderCompilerTraversers
                 }
             }
 
-            static ScopeT Traverse(TProgram& program, IdGenerator& ids)
-            {
-                auto* scope = new AllocationsScope();
-                Traverse(program.getIntermediate(EShLangVertex), ids, *scope);
-                Traverse(program.getIntermediate(EShLangFragment), ids, *scope);
-                return std::unique_ptr<AllocationsScopeBase>(scope);
-            }
-
-        private:
-            UniformTypeChangeTraverser(TIntermediate* intermediate, AllocationsScope& scope)
-                : TIntermTraverser{}
-                , m_intermediate{intermediate}
-                , m_scope{scope}
-            {
-            }
-
-            static void Traverse(glslang::TIntermediate* intermediate, IdGenerator&, AllocationsScope& scope)
+            static void Traverse(TIntermediate* intermediate, IdGenerator&, AllocationsScope& scope)
             {
                 UniformTypeChangeTraverser traverser{intermediate, scope};
                 intermediate->getTreeRoot()->traverse(&traverser);
@@ -448,19 +413,19 @@ namespace Babylon::ShaderCompilerTraversers
             AllocationsScope& m_scope;
         };
 
-        /// <summary>
         /// This traverser modifies all vertex attributes (position, UV, etc.) to conform to
         /// bgfx's expectations regarding name and location. It is currently required for
         /// DirectX, OpenGL, and Metal.
-        /// </summary>
         class VertexVaryingInTraverser final : private TIntermTraverser
         {
         public:
-            ~VertexVaryingInTraverser() override
+            static void Traverse(TProgram& program, IdGenerator& ids)
             {
+                Traverse(program.getIntermediate(EShLangVertex), ids);
             }
 
-            void visitSymbol(TIntermSymbol* symbol) override
+        private:
+            virtual void visitSymbol(TIntermSymbol* symbol) override
             {
                 // Collect all vertex attributes, described by glslang as "varyings."
                 if (symbol->getType().getQualifier().storage == EvqVaryingIn)
@@ -481,12 +446,6 @@ namespace Babylon::ShaderCompilerTraversers
                 }
             }
 
-            static void Traverse(TProgram& program, IdGenerator& ids)
-            {
-                Traverse(program.getIntermediate(EShLangVertex), ids);
-            }
-
-        private:
             std::pair<unsigned int, const char*> GetVaryingLocationAndNewNameForName(const char* name)
             {
 #define IF_NAME_RETURN_ATTRIB(varyingName, attrib, newName)  \
@@ -508,7 +467,7 @@ namespace Babylon::ShaderCompilerTraversers
                 return {FIRST_GENERIC_ATTRIBUTE_LOCATION + m_genericAttributesRunningCount++, name};
             }
 
-            static void Traverse(glslang::TIntermediate* intermediate, IdGenerator& ids)
+            static void Traverse(TIntermediate* intermediate, IdGenerator& ids)
             {
                 VertexVaryingInTraverser traverser{};
                 intermediate->getTreeRoot()->traverse(&traverser);
@@ -576,10 +535,6 @@ namespace Babylon::ShaderCompilerTraversers
         class SamplerSplitterTraverser final : TIntermTraverser
         {
         public:
-            ~SamplerSplitterTraverser() override
-            {
-            }
-
             void visitSymbol(TIntermSymbol* symbol) override
             {
                 if (symbol->getType().getQualifier().storage == EvqUniform && symbol->getType().getBasicType() == EbtSampler)
@@ -710,6 +665,41 @@ namespace Babylon::ShaderCompilerTraversers
             std::map<std::string, TIntermSymbol*> m_samplerNameToSymbol{};
             std::vector<std::pair<TIntermSymbol*, TIntermNode*>> m_symbolsToParents{};
         };
+
+        class InvertYDerivativeOperandsTraverser : public TIntermTraverser
+        {
+        public:
+            static void Traverse(TProgram& program)
+            {
+                auto intermediate{program.getIntermediate(EShLangFragment)};
+                InvertYDerivativeOperandsTraverser invertYDerivativeOperandsTraverser{intermediate};
+                intermediate->getTreeRoot()->traverse(&invertYDerivativeOperandsTraverser);
+            }
+
+        protected:
+            virtual bool visitUnary(TVisit visit, TIntermUnary* unary) override
+            {
+                if (visit == EvPreVisit)
+                {
+                    auto op = unary->getOp();
+                    if (op == EOpDPdy || op == EOpDPdyFine || op == EOpDPdyCoarse)
+                    {
+                        unary->setOperand(m_intermediate->addUnaryNode(EOpNegative, unary->getOperand(), {}));
+                        return false;
+                    }
+                }
+
+                return true;
+            }
+
+        private:
+            InvertYDerivativeOperandsTraverser(TIntermediate* intermediate)
+                : m_intermediate{intermediate}
+            {
+            }
+
+            TIntermediate* m_intermediate;
+        };
     }
 
     ScopeT MoveNonSamplerUniformsIntoStruct(TProgram& program, IdGenerator& ids)
@@ -730,5 +720,10 @@ namespace Babylon::ShaderCompilerTraversers
     void SplitSamplersIntoSamplersAndTextures(TProgram& program, IdGenerator& ids)
     {
         SamplerSplitterTraverser::Traverse(program, ids);
+    }
+
+    void InvertYDerivativeOperands(TProgram& program)
+    {
+        InvertYDerivativeOperandsTraverser::Traverse(program);
     }
 }
